# schema/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Bar": {
        "description": "An aggregated OHLCV bar.\n\n``` use honba_messages::{ Bar, BarAggregation, BarSpecification, BarType, InstrumentId, PriceType, UnixNanos, Exchange, };\n\nlet bar_type = BarType::new( InstrumentId::new(\"NIFTY50\", Exchange::new(\"NSE\")), BarSpecification::new(1, BarAggregation::Minute, PriceType::Last), ); let bar = Bar::new( bar_type, 22_000.0, 22_050.0, 21_980.0, 22_020.0, 15_000.0, UnixNanos::from_u64(1_700_000_060_000_000_000), UnixNanos::from_u64(1_700_000_060_000_000_000), ); assert_eq!(bar.open(), 22_000.0); assert_eq!(bar.close(), 22_020.0); ```\n\nInvariants (checked by [`Bar::validate`] and on deserialization): all prices and the volume are finite, `low <= open, close <= high`, and `volume >= 0`.",
        "properties": {
          "bar_type": {
            "$ref": "#/components/schemas/BarType"
          },
          "close": {
            "format": "double",
            "type": "number"
          },
          "high": {
            "format": "double",
            "type": "number"
          },
          "low": {
            "format": "double",
            "type": "number"
          },
          "open": {
            "format": "double",
            "type": "number"
          },
          "ts_event": {
            "$ref": "#/components/schemas/UnixNanos"
          },
          "ts_init": {
            "$ref": "#/components/schemas/UnixNanos"
          },
          "volume": {
            "format": "double",
            "type": "number"
          }
        },
        "required": [
          "bar_type",
          "close",
          "high",
          "low",
          "open",
          "ts_event",
          "ts_init",
          "volume"
        ],
        "type": "object"
      },
      "BarAggregation": {
        "description": "How a bar aggregates its underlying data.",
        "oneOf": [
          {
            "description": "Aggregate every N ticks.",
            "enum": [
              "tick"
            ],
            "type": "string"
          },
          {
            "description": "Aggregate every N seconds.",
            "enum": [
              "second"
            ],
            "type": "string"
          },
          {
            "description": "Aggregate every N minutes.",
            "enum": [
              "minute"
            ],
            "type": "string"
          },
          {
            "description": "Aggregate every N hours.",
            "enum": [
              "hour"
            ],
            "type": "string"
          },
          {
            "description": "Aggregate every N days.",
            "enum": [
              "day"
            ],
            "type": "string"
          },
          {
            "description": "Aggregate every N weeks.",
            "enum": [
              "week"
            ],
            "type": "string"
          },
          {
            "description": "Aggregate every N calendar months.",
            "enum": [
              "month"
            ],
            "type": "string"
          }
        ]
      },
      "BarSpecification": {
        "description": "Describes how a [`Bar`] aggregates its inputs.\n\n``` use honba_messages::{BarAggregation, BarSpecification, PriceType};\n\nlet spec = BarSpecification::new(1, BarAggregation::Minute, PriceType::Last); assert_eq!(spec.step(), 1); assert_eq!(spec.aggregation(), BarAggregation::Minute); ```\n\n`step` must be at least 1; deserializing a zero step fails.",
        "properties": {
          "aggregation": {
            "$ref": "#/components/schemas/BarAggregation"
          },
          "price_type": {
            "$ref": "#/components/schemas/PriceType"
          },
          "step": {
            "format": "uint",
            "minimum": 0.0,
            "type": "integer"
          }
        },
        "required": [
          "aggregation",
          "price_type",
          "step"
        ],
        "type": "object"
      },
      "BarType": {
        "description": "Fully identifies a bar: which instrument, and how it aggregates.",
        "properties": {
          "instrument_id": {
            "$ref": "#/components/schemas/InstrumentId"
          },
          "spec": {
            "$ref": "#/components/schemas/BarSpecification"
          }
        },
        "required": [
          "instrument_id",
          "spec"
        ],
        "type": "object"
      },
...
      "BarsResponse": {
        "description": "Bars response.",
        "properties": {
          "bars": {
            "description": "Bars in ascending `ts_event` order.",
            "items": {
              "$ref": "#/components/schemas/Bar"
            },
            "type": "array"
          }
        },
        "required": [
          "bars"
        ],
        "type": "object"
      },
...
      "DepthLevel": {
        "description": "One side of the order book.",
        "properties": {
          "price": {
            "description": "Price.",
            "format": "double",
            "type": "number"
          },
          "qty": {
            "description": "Quantity resting at this price.",
            "format": "double",
            "type": "number"
          }
        },
        "required": [
          "price",
          "qty"
        ],
        "type": "object"
      },
...
      "DepthResponse": {
        "description": "Depth (order book) response.",
        "properties": {
          "asks": {
            "description": "Ask levels, best (lowest) first.",
            "items": {
              "$ref": "#/components/schemas/DepthLevel"
            },
            "type": "array"
          },
          "bids": {
            "description": "Bid levels, best (highest) first.",
            "items": {
              "$ref": "#/components/schemas/DepthLevel"
            },
            "type": "array"
          }
        },
        "required": [
          "asks",
          "bids"
        ],
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "One code per failure mode, stable across releases (plan.md §4.2).\n\nDeserialization is deliberately strict: an unrecognized code is an error rather than a default, because silently mapping a newer server's code onto a known one would report the wrong thing happened.\n\n``` use honba_messages::ErrorCode;\n\nlet code = ErrorCode::RiskMaxNotionalExceeded; assert_eq!(code.category().as_str(), \"risk\"); assert!(!code.is_retryable(), \"a risk refusal is not worth repeating\"); assert_eq!( serde_json::to_string(&code).unwrap(), \"\\\"risk_max_notional_exceeded\\\"\" ); ```",
        "oneOf": [
          {
            "description": "A request was malformed or failed validation.",
            "enum": [
              "validation_invalid_request"
            ],
            "type": "string"
          },
          {
            "description": "A named resource does not exist.",
            "enum": [
              "not_found"
            ],
            "type": "string"
          },
          {
            "description": "The caller is not authenticated.",
            "enum": [
              "unauthorized"
            ],
            "type": "string"
          },
          {
            "description": "The caller is authenticated but not permitted to do this.",
            "enum": [
              "forbidden"
            ],
            "type": "string"
          },
          {
            "description": "The caller exceeded a rate limit.",
            "enum": [
              "rate_limited"
            ],
            "type": "string"
          },
          {
            "description": "Order notional exceeded the configured maximum.",
            "enum": [
              "risk_max_notional_exceeded"
            ],
            "type": "string"
          },
          {
            "description": "Position size exceeded the configured maximum.",
            "enum": [
              "risk_max_position_exceeded"
            ],
            "type": "string"
          },
          {
            "description": "Account drawdown exceeded the configured maximum.",
            "enum": [
              "risk_max_drawdown_exceeded"
            ],
            "type": "string"
          },
          {
            "description": "The exchange or gateway rejected the order.",
            "enum": [
              "order_rejected"
            ],
            "type": "string"
          },
          {
            "description": "The referenced order does not exist.",
            "enum": [
              "order_not_found"
            ],
            "type": "string"
          },
          {
            "description": "The referenced instrument does not exist.",
            "enum": [
              "instrument_not_found"
            ],
            "type": "string"
          },
          {
            "description": "Market data is unavailable for the request.",
            "enum": [
              "market_data_unavailable"
            ],
            "type": "string"
          },
          {
            "description": "The operation timed out.",
            "enum": [
              "timeout"
            ],
            "type": "string"
          },
          {
            "description": "A transport or network failure.",
            "enum": [
              "transport_error"
            ],
            "type": "string"
          },
          {
            "description": "An unanticipated server-side failure.",
            "enum": [
              "internal_error"
            ],
            "type": "string"
          },
          {
            "description": "The operation is not supported on this surface or build.",
            "enum": [
              "unsupported"
            ],
            "type": "string"
          },
          {
            "description": "The route exists in the contract but its behaviour is not built yet.",
            "enum": [
              "not_implemented"
            ],
            "type": "string"
          }
        ]
      },
      "ErrorDetail": {
        "description": "A single error as it crosses a surface.\n\nCarries the stable [`ErrorCode`] plus human-readable detail. `context` holds structured specifics (the limit that was exceeded, the id that was not found) so a client can react programmatically without parsing `message`.",
        "properties": {
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ],
            "description": "The stable machine-readable code."
          },
          "context": {
            "description": "Structured detail specific to this failure."
          },
          "message": {
            "description": "Human-readable description. Never parsed by clients.",
            "type": "string"
          },
          "retryable": {
            "description": "Whether repeating the identical operation is safe.",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "type": "object"
      },
...
      "Exchange": {
        "description": "A trading exchange, e.g. `NSE`, `BSE`, `NASDAQ`.",
        "type": "string"
      },
...
      "InstrumentId": {
        "description": "A trading instrument, combining a symbol and its exchange.",
        "properties": {
          "exchange": {
            "$ref": "#/components/schemas/Exchange"
          },
          "symbol": {
            "type": "string"
          }
        },
        "required": [
          "exchange",
          "symbol"
        ],
        "type": "object"
      },
...
      "InstrumentsResponse": {
        "description": "Instruments response.",
        "properties": {
          "instruments": {
            "description": "Matching instruments, as of the requested snapshot.",
            "items": true,
            "type": "array"
          }
        },
        "required": [
          "instruments"
        ],
        "type": "object"
      },
...
      "PriceType": {
        "description": "Which price of the underlying data feeds the bar.",
        "oneOf": [
          {
            "description": "Use the bid price.",
            "enum": [
              "bid"
            ],
            "type": "string"
          },
          {
            "description": "Use the ask price.",
            "enum": [
              "ask"
            ],
            "type": "string"
          },
          {
            "description": "Use the midpoint.",
            "enum": [
              "mid"
            ],
            "type": "string"
          },
          {
            "description": "Use the last traded price.",
            "enum": [
              "last"
            ],
            "type": "string"
          }
        ]
      },
      "QuoteTick": {
        "description": "A top-of-book quote update.\n\n``` use honba_messages::{InstrumentId, QuoteTick, UnixNanos, Exchange};\n\nlet tick = QuoteTick::new( InstrumentId::new(\"NIFTY50\", Exchange::new(\"NSE\")), 22_000.0, 22_001.0, 50.0, 75.0, UnixNanos::from_u64(1), UnixNanos::from_u64(1), ); assert_eq!(tick.bid_price(), 22_000.0); assert_eq!(tick.ask_price(), 22_001.0); ```\n\nInvariants (checked by [`QuoteTick::validate`] and on deserialization): prices and sizes are finite, `bid_price <= ask_price`, sizes are `>= 0`.",
        "properties": {
          "ask_price": {
            "format": "double",
            "type": "number"
          },
          "ask_size": {
            "format": "double",
            "type": "number"
          },
          "bid_price": {
            "format": "double",
            "type": "number"
          },
          "bid_size": {
            "format": "double",
            "type": "number"
          },
          "instrument_id": {
            "$ref": "#/components/schemas/InstrumentId"
          },
          "ts_event": {
            "$ref": "#/components/schemas/UnixNanos"
          },
          "ts_init": {
            "$ref": "#/components/schemas/UnixNanos"
          }
        },
        "required": [
          "ask_price",
          "ask_size",
          "bid_price",
          "bid_size",
          "instrument_id",
          "ts_event",
          "ts_init"
        ],
        "type": "object"
      },
...
      "QuotesResponse": {
        "description": "Quotes response.",
        "properties": {
          "quotes": {
            "description": "Latest top-of-book per requested symbol.",
            "items": {
              "$ref": "#/components/schemas/QuoteTick"
            },
            "type": "array"
          }
        },
        "required": [
          "quotes"
        ],
        "type": "object"
      },
      "ResponseEnvelope": {
        "description": "A response body, always carrying both version axes and at most one of `data` / `error`.\n\nUnknown fields are accepted on read so a v1 reader can consume a response from a newer writer that added an optional field; unknown *versions* are rejected by the caller, which compares `api_version` and `schema_version` against what it was built for.",
        "properties": {
          "api_version": {
            "description": "Semantic version of the API surface.",
            "type": "string"
          },
          "data": {
            "description": "Payload on success; absent on failure.",
            "type": "null"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error detail on failure; absent on success."
          },
          "schema_version": {
            "description": "Integer version of the wire shape (see [`SCHEMA_VERSION`](crate::SCHEMA_VERSION)).",
            "format": "uint32",
            "minimum": 0.0,
            "type": "integer"
          }
        },
        "required": [
          "api_version",
          "schema_version"
        ],
        "type": "object"
      },
...
      "UnixNanos": {
        "description": "JSON representation of UnixNanos with ISO-8601 string and unix_nanos string.",
        "properties": {
          "iso": {
            "type": "string"
          },
          "unix_nanos": {
            "type": "string"
          }
        },
        "required": [
          "iso",
          "unix_nanos"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Generated by honba-codegen from the Rust source of truth. DO NOT EDIT: regenerate with `make openapi`.",
    "title": "Honba API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/bars/{id}": {
      "get": {
        "operationId": "get_bars_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "description": "Inclusive start time, RFC3339 or ISO-8601 date.",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "tf",
            "required": false,
            "schema": {
              "description": "Timeframe/bar specification (e.g. \"1m\", \"1d\").",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "description": "Exclusive end time, RFC3339 or ISO-8601 date.",
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarsResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET bars  id",
        "tags": [
          "bars"
        ],
        "x-honba-access": "read_only"
      }
    },
...
    "/depth/{id}": {
      "get": {
        "operationId": "get_depth_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "description": "Number of levels per side.",
              "format": "uint32",
              "minimum": 0.0,
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepthResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET depth  id",
        "tags": [
          "depth"
        ],
        "x-honba-access": "read_only"
      }
    },
...
    "/instruments": {
      "get": {
        "operationId": "get_instruments",
        "parameters": [
          {
            "in": "query",
            "name": "exchange",
            "required": false,
            "schema": {
              "description": "Exchange filter.",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "description": "Symbol filter.",
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentsResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET instruments",
        "tags": [
          "instruments"
        ],
        "x-honba-access": "read_only"
      }
    },
    "/instruments/{id}": {
      "get": {
        "operationId": "get_instruments_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentsResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET instruments  id",
        "tags": [
          "instruments"
        ],
        "x-honba-access": "read_only"
      }
    },
...
    "/quotes": {
      "get": {
        "operationId": "get_quotes",
        "parameters": [
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "description": "Return the latest quote known at this time (RFC3339 or ISO-8601 date, inclusive); the latest held when omitted.",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "symbols",
            "required": false,
            "schema": {
              "description": "Comma-separated symbols.",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "venue",
            "required": false,
            "schema": {
              "description": "Venue/exchange.",
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotesResponse"
                }
              }
            },
            "description": "Success. Always enveloped; see ResponseEnvelope."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Failure. Carries the envelope's `error` with a stable ErrorCode."
          }
        },
        "summary": "GET quotes",
        "tags": [
          "quotes"
        ],
        "x-honba-access": "read_only"
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}